clean up and leftovers from content as data PR
